+++ apps/filecoin-site/src/app/[locale]/case-studies/[slug]/utils/generateStructuredData.ts
@@ -1,15 +1,17 @@
 import { PATHS } from '@/constants/paths'
 
 import type { CaseStudy } from '../../types/caseStudyType'
-import { generateCaseStudyArticleStructuredData } from '../../utils/generateCaseStudyArticleStructuredData'
+import { generateCaseStudyPageStructuredData } from '../../utils/generateCaseStudyPageStructuredData'
 
 export function generateStructuredData(data: CaseStudy) {
-  const { title, pageDescription, image } = data
+  const { seo, image, publishedOn, updatedOn } = data
 
-  return generateCaseStudyArticleStructuredData({
+  return generateCaseStudyPageStructuredData({
     path: `${PATHS.CASE_STUDIES.path}/${data.slug}`,
-    headline: title,
-    description: pageDescription,
-    image: image ? image.src : undefined, // #todo: Make image required
+    headline: seo.title,
+    description: seo.description,
+    image: image?.src || '',
+    datePublished: publishedOn,
+    dateModified: updatedOn,
   })
 }